requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
langchain
langchain-openai
pydantic
openai
typing-extensions
python-multipart
requests
pydub
#audioop-lts